requirements.txt
cryptography==41.0.7
matplotlib==3.8.0
mysql-connector-python==8.2.0
openpyxl==3.1.2
pandas==2.1.3
plotly==5.18.0
pycountry==22.3.5
reportlab==3.6.13
requests==2.31.0
scikit-learn==1.3.2
seaborn==0.13.0
tabulate==0.9.0
